LayoutTests/fast/css/stylesheet-parameters-expected.txt

CONSOLE MESSAGE: line 1: Calling CSSStyleSheet.insertRule() with one argument is deprecated. Please pass the index argument as well: insertRule(x, 0).
Tests that the parameters to CSSStyleSheet.insertRule() / deleteRule() are mandatory.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS stylesheet.__proto__ is CSSStyleSheet.prototype
PASS stylesheet.deleteRule() threw exception TypeError: Not enough arguments.
PASS stylesheet.insertRule() threw exception TypeError: Not enough arguments.
FAIL stylesheet.insertRule('body { margin: 0; }') should throw TypeError: Not enough arguments. Was 0.
PASS successfullyParsed is true

TEST COMPLETE
